No Commit for Demo Users

+++ app/Http/Controllers/AdminCampsController.php
@@ -64,16 +64,19 @@ class AdminCampsController extends Controller
     {
         //
         $input = $request->all();
-        $user = User::findOrFail(Auth::user()->id);
 
-        if (! $user->isAdmin()) {
-            $input['user_id'] = $user->id;
-        }
+        if (!Auth::user()->demo) {
+            $user = User::findOrFail(Auth::user()->id);
 
-        $camp = Camp::create($input);
-        CampCreated::dispatch($camp);
-        if (! $user->isAdmin()) {
-            $user->update(['camp_id' => $camp->id]);
+            if (!$user->isAdmin()) {
+                $input['user_id'] = $user->id;
+            }
+
+            $camp = Camp::create($input);
+            CampCreated::dispatch($camp);
+            if (!$user->isAdmin()) {
+                $user->update(['camp_id' => $camp->id]);
+            }
         }
 
         return redirect('admin/camps');
@@ -116,7 +119,7 @@ class AdminCampsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if (! Auth::user()->demo) {
+        if (!Auth::user()->demo) {
             Camp::findOrFail($id)->update($request->all());
         }
 
@@ -131,30 +134,34 @@ class AdminCampsController extends Controller
      */
     public function destroy(Camp $camp)
     {
-        $users = Auth::user()->camp->allUsers;
-        $camp_global = Camp::where('global_camp', true)->first();
+        if (!Auth::user()->demo) {
+            $users = Auth::user()->camp->allUsers;
+            $camp_global = Camp::where('global_camp', true)->first();
 
-        foreach ($users as $user) {
-            Helper::updateCamp($user, $camp_global);
+            foreach ($users as $user) {
+                Helper::updateCamp($user, $camp_global);
+            }
+            $counter = $camp->surveys()->count();
+            foreach ($camp->camp_users_all()->get() as $camp_user) {
+                $camp_user->delete();
+            }
+            foreach ($camp->posts()->get() as $post) {
+                $post->delete();
+            }
+            File::deleteDirectory(storage_path('app/public/' . Str::slug($camp['name'])));
+            File::deleteDirectory(storage_path('app/files/' . Str::slug($camp['name'])));
+            $camp->update(['finish' => true, 'counter' => $counter]);
         }
-        $counter = $camp->surveys()->count();
-        foreach ($camp->camp_users_all()->get() as $camp_user) {
-            $camp_user->delete();
-        }
-        foreach ($camp->posts()->get() as $post) {
-            $post->delete();
-        }
-        File::deleteDirectory(storage_path('app/public/'.Str::slug($camp['name'])));
-        File::deleteDirectory(storage_path('app/files/'.Str::slug($camp['name'])));
-        $camp->update(['finish' => true, 'counter' => $counter]);
 
         return redirect('/home');
     }
 
     public function opensurvey()
     {
-        $camp = Auth::user()->camp;
-        $camp->update(['secondsurveyopen' => true]);
+        if (! Auth::user()->demo) {
+            $camp = Auth::user()->camp;
+            $camp->update(['secondsurveyopen' => true]);
+        }
 
         return redirect('/admin/surveys');
     }
