+++ app/Http/Controllers/AdminUsersController.php
@@ -100,6 +100,7 @@ class AdminUsersController extends Controller
     public function create()
     {
         //
+
         $aktUser = Auth::user();
         if ($aktUser->isAdmin()) {
             $roles = Role::pluck('name', 'id')->all();
@@ -119,165 +120,171 @@ class AdminUsersController extends Controller
 
     public function import()
     {
-        $aktUser = Auth::user();
-        $camp = $aktUser->camp;
-        if ($aktUser->foreign_id && $camp->foreign_id && $camp->group && $camp->group['api_token']) {
-            $url = 'https://db.cevi.ch/groups/'.$camp->group['foreign_id'].'/events/'.$camp['foreign_id'].'/participations.json';
-            $response = Curl::to( $url)
-                ->withData(['token' => Crypt::decryptString($camp->group['api_token'])])
-                ->get();
-            $response = json_decode($response);
-            $participants = $response->event_participations;
-            for ($i = 1; $i < $response->total_pages; $i++) {
-                $url = $response->next_page_link;
+
+        if (!Auth::user()->demo) {
+            $aktUser = Auth::user();
+            $camp = $aktUser->camp;
+            if ($aktUser->foreign_id && $camp->foreign_id && $camp->group && $camp->group['api_token']) {
+                $url = 'https://db.cevi.ch/groups/' . $camp->group['foreign_id'] . '/events/' . $camp['foreign_id'] . '/participations.json';
                 $response = Curl::to($url)
-                    ->withData([
-                        'page' => $i + 1,
-                        'token' => Crypt::decryptString($camp->group['api_token'])
-                    ])
+                    ->withData(['token' => Crypt::decryptString($camp->group['api_token'])])
                     ->get();
                 $response = json_decode($response);
-                $participants = array_merge($participants, $response->event_participations);
-            }
-            $isLeader = false;
-            foreach ($participants as $participant) {
-                if ((intval($participant->links->person) === $aktUser['foreign_id']) && ($participant->roles[0]->type === 'Event::Role::Leader')) {
-                    $isLeader = true;
+                $participants = $response->event_participations;
+                for ($i = 1; $i < $response->total_pages; $i++) {
+                    $url = $response->next_page_link;
+                    $response = Curl::to($url)
+                        ->withData([
+                            'page' => $i + 1,
+                            'token' => Crypt::decryptString($camp->group['api_token'])
+                        ])
+                        ->get();
+                    $response = json_decode($response);
+                    $participants = array_merge($participants, $response->event_participations);
                 }
-            }
-            if ($isLeader) {
+                $isLeader = false;
                 foreach ($participants as $participant) {
-                    if ($participant->roles[0]->type === 'Event::Course::Role::Participant' ||
+                    if ((intval($participant->links->person) === $aktUser['foreign_id']) && ($participant->roles[0]->type === 'Event::Role::Leader')) {
+                        $isLeader = true;
+                    }
+                }
+                if ($isLeader) {
+                    foreach ($participants as $participant) {
+                        if ($participant->roles[0]->type === 'Event::Course::Role::Participant' ||
                             $participant->roles[0]->type === 'Event::Role::AssistantLeader' ||
                             $participant->roles[0]->type === 'Event::Role::Leader') {
-                        if ($participant->links->person != $aktUser['foreign_id']) {
-                            $username = $participant->nickname ? $participant->nickname : $participant->first_name;
-                            switch($participant->roles[0]->type) {
-                                case  'Event::Course::Role::Participant':
-                                    $role_id = config('status.role_Teilnehmer');
-                                    break;
-                                case  'Event::Role::AssistantLeader':
-                                    $role_id = config('status.role_Gruppenleiter');
-                                    break;
-                                case  'Event::Role::Leader':
-                                    $role_id = config('status.role_Kursleiter');
-                                    break;
+                            if ($participant->links->person != $aktUser['foreign_id']) {
+                                $username = $participant->nickname ? $participant->nickname : $participant->first_name;
+                                switch ($participant->roles[0]->type) {
+                                    case  'Event::Course::Role::Participant':
+                                        $role_id = config('status.role_Teilnehmer');
+                                        break;
+                                    case  'Event::Role::AssistantLeader':
+                                        $role_id = config('status.role_Gruppenleiter');
+                                        break;
+                                    case  'Event::Role::Leader':
+                                        $role_id = config('status.role_Kursleiter');
+                                        break;
+                                }
+                                $insertData = [
+                                    'username' => $username,
+                                    'email' => $participant->email,
+                                    'email_verified_at' => now(),
+                                    'classification_id' => config('status.classification_yellow'),];
+                                $user = User::whereraw('LOWER(`email`) LIKE "' . mb_strtolower($participant->email) . '"')->Orwhere('foreign_id', $participant->links->person)->first();
+                                if (!$user) {
+                                    $user = User::create($insertData);
+                                    UserCreated::dispatch($user);
+                                }
+                                $user->update([
+                                    'role_id' => $role_id,
+                                    'camp_id' => $camp['id'],
+                                    'classification_id' => config('status.classification_yellow'),
+                                ]);
+                            } else {
+                                $user = Auth::user();
                             }
-                            $insertData = [
-                                'username' => $username,
-                                'email' => $participant->email,
-                                'email_verified_at' => now(),
-                                'classification_id' => config('status.classification_yellow'), ];
-                            $user = User::whereraw('LOWER(`email`) LIKE "'.mb_strtolower($participant->email).'"')->Orwhere('foreign_id', $participant->links->person)->first();
-                            if (! $user) {
-                                $user = User::create($insertData);
-                                UserCreated::dispatch($user);
-                            }
-                            $user->update([
-                                'role_id' => $role_id,
-                                'camp_id' => $camp['id'],
-                                'classification_id' => config('status.classification_yellow'),
-                            ]);
-                        } else {
-                            $user = Auth::user();
-                        }
 //                        if(!$user->avatar){
 //                            $user->update(['avatar' => 'https://db.cevi.ch'. $participant->picture->url]);
 //                        }
-                        if (! $user->email) {
-                            $user->update(['email' => $participant->email]);
+                            if (!$user->email) {
+                                $user->update(['email' => $participant->email]);
+                            }
+                            if (!$user->foreign_id) {
+                                $user->update(['foreign_id' => $participant->links->person]);
+                            }
+                            Helper::updateCamp($user, $camp);
                         }
-                        if (! $user->foreign_id) {
-                            $user->update(['foreign_id' => $participant->links->person]);
-                        }
-                        Helper::updateCamp($user, $camp);
                     }
-                }
 
-                return true;
+                    return true;
+                } else {
+                    $errorText = 'Der DB-Import steht nur den Kursleitern zur Verfügung';
+                    abort(412, $errorText);
+                }
             } else {
-                $errorText = 'Der DB-Import steht nur den Kursleitern zur Verfügung';
+                $errorText = '';
+                if (!$camp->foreign_id) {
+                    $errorText = 'Keine Cevi-DB-ID auf dem Kurs hinterlegt.';
+                }
+                if (!$aktUser->foreign_id) {
+                    $errorText = 'Dein Cevi-DB-Benutzer ist noch nicht mit deinem Benutzer verknüpft.';
+                } elseif (!$camp->group) {
+                    $errorText = $errorText + ' Keine Gruppe auf dem Kurs hinterlegt.';
+                } elseif (!$camp->group['api_token']) {
+                    $errorText = $errorText + ' Deine Region hat den DB-Import nicht freigeschalten.';
+                }
                 abort(412, $errorText);
             }
-        } else {
-            $errorText = '';
-            if (! $camp->foreign_id) {
-                $errorText = 'Keine Cevi-DB-ID auf dem Kurs hinterlegt.';
-            }
-            if (! $aktUser->foreign_id) {
-                $errorText = 'Dein Cevi-DB-Benutzer ist noch nicht mit deinem Benutzer verknüpft.';
-            } elseif (! $camp->group) {
-                $errorText = $errorText + ' Keine Gruppe auf dem Kurs hinterlegt.';
-            } elseif (! $camp->group['api_token']) {
-                $errorText = $errorText + ' Deine Region hat den DB-Import nicht freigeschalten.';
-            }
-            abort(412, $errorText);
         }
     }
 
     public function uploadFile(Request $request)
     {
-        if ($request->hasFile('csv_file')) {
-            $array = (new UsersImport)->toArray(request()->file('csv_file'));
-            $importData_arr = $array[0];
 
-            // Insert to MySQL database
-            $user = Auth::user();
-            $camp = $user->camp;
-            foreach ($importData_arr as $importData) {
-                $username = mb_strtolower($importData['ceviname']);
+        if (!Auth::user()->demo) {
+            if ($request->hasFile('csv_file')) {
+                $array = (new UsersImport)->toArray(request()->file('csv_file'));
+                $importData_arr = $array[0];
 
-                if ($importData['rollen'] === 'K') {
-                    $insertData = [
+                // Insert to MySQL database
+                $user = Auth::user();
+                $camp = $user->camp;
+                foreach ($importData_arr as $importData) {
+                    $username = mb_strtolower($importData['ceviname']);
 
-                        'username' => $username,
-                        'email' => $importData['e_mail'],
-                        'password' => bcrypt($importData['passwort']),
-                        'role_id' => config('status.role_Kursleiter'),
-                        'camp_id' => $user['camp_id'],
-                        'email_verified_at' => now(),
-                        'classification_id' => config('status.classification_yellow'), ];
+                    if ($importData['rollen'] === 'K') {
+                        $insertData = [
 
-                    $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
-                    UserCreated::dispatch($user);
-                    Helper::updateCamp($user, $camp);
-                } elseif ($importData['rollen'] === 'G') {
-                    $insertData = [
+                            'username' => $username,
+                            'email' => $importData['e_mail'],
+                            'password' => bcrypt($importData['passwort']),
+                            'role_id' => config('status.role_Kursleiter'),
+                            'camp_id' => $user['camp_id'],
+                            'email_verified_at' => now(),
+                            'classification_id' => config('status.classification_yellow'),];
 
-                        'username' => $username,
-                        'email' => $importData['e_mail'],
-                        'password' => bcrypt($importData['passwort']),
-                        'role_id' => config('status.role_Gruppenleiter'),
-                        'camp_id' => $user['camp_id'],
-                        'email_verified_at' => now(),
-                        'classification_id' => config('status.classification_yellow'), ];
+                        $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
+                        UserCreated::dispatch($user);
+                        Helper::updateCamp($user, $camp);
+                    } elseif ($importData['rollen'] === 'G') {
+                        $insertData = [
 
-                    $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
-                    UserCreated::dispatch($user);
-                    Helper::updateCamp($user, $camp);
+                            'username' => $username,
+                            'email' => $importData['e_mail'],
+                            'password' => bcrypt($importData['passwort']),
+                            'role_id' => config('status.role_Gruppenleiter'),
+                            'camp_id' => $user['camp_id'],
+                            'email_verified_at' => now(),
+                            'classification_id' => config('status.classification_yellow'),];
+
+                        $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
+                        UserCreated::dispatch($user);
+                        Helper::updateCamp($user, $camp);
+                    }
                 }
-            }
-            foreach ($importData_arr as $importData) {
-                $username = mb_strtolower($importData['ceviname']);
+                foreach ($importData_arr as $importData) {
+                    $username = mb_strtolower($importData['ceviname']);
 
-                if ($importData['rollen'] === 'T') {
-                    $user = Auth::user();
-                    $leader = User::where('email', $importData['leiter'])->first();
+                    if ($importData['rollen'] === 'T') {
+                        $user = Auth::user();
+                        $leader = User::where('email', $importData['leiter'])->first();
 
-                    $insertData = [
+                        $insertData = [
 
-                        'username' => $username,
-                        'email' => $importData['e_mail'],
-                        'password' => bcrypt($importData['passwort']),
-                        'role_id' => config('status.role_Teilnehmer'),
-                        'camp_id' => $user['camp_id'],
-                        'email_verified_at' => now(),
-                        'leader_id' => $leader['id'],
-                        'classification_id' => config('status.classification_yellow'), ];
+                            'username' => $username,
+                            'email' => $importData['e_mail'],
+                            'password' => bcrypt($importData['passwort']),
+                            'role_id' => config('status.role_Teilnehmer'),
+                            'camp_id' => $user['camp_id'],
+                            'email_verified_at' => now(),
+                            'leader_id' => $leader['id'],
+                            'classification_id' => config('status.classification_yellow'),];
 
-                    $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
-                    UserCreated::dispatch($user);
-                    Helper::updateCamp($user, $camp);
+                        $user = User::firstOrCreate(['email' => $insertData['email']], $insertData);
+                        UserCreated::dispatch($user);
+                        Helper::updateCamp($user, $camp);
+                    }
                 }
             }
         }
@@ -297,24 +304,27 @@ class AdminUsersController extends Controller
             'role_id.required' => 'Die Rolle muss ausgefüllt sein.', ]);
 
         $aktUser = Auth::user();
-        $camp = $aktUser->camp;
-        if (trim($request->password) == '') {
-            $input = $request->except('password');
-        } else {
-            $input = $request->all();
-            $input['password'] = bcrypt($request->password);
-        }
 
-        if (! $aktUser->isAdmin()) {
-            $input['camp_id'] = $camp['id'];
-        }
-        // $input['slug'] = Str::slug($input['username']);
-        $input['email_verified_at'] = now();
+        if (!Auth::user()->demo) {
+            $camp = $aktUser->camp;
+            if (trim($request->password) == '') {
+                $input = $request->except('password');
+            } else {
+                $input = $request->all();
+                $input['password'] = bcrypt($request->password);
+            }
 
-        $user = User::create($input);
-        UserCreated::dispatch($user);
-        Helper::updateCamp($user, $camp);
-        Helper::updateAvatar($request, $user);
+            if (!$aktUser->isAdmin()) {
+                $input['camp_id'] = $camp['id'];
+            }
+            // $input['slug'] = Str::slug($input['username']);
+            $input['email_verified_at'] = now();
+
+            $user = User::create($input);
+            UserCreated::dispatch($user);
+            Helper::updateCamp($user, $camp);
+            Helper::updateAvatar($request, $user);
+        }
 
         return redirect('/admin/users/create');
     }
@@ -322,15 +332,18 @@ class AdminUsersController extends Controller
     public function add(Request $request)
     {
         $input = $request->all();
-        if ($input['user_id']) {
-            $aktUser = Auth::user();
-            $camp = $aktUser->camp;
-            $user = User::findOrFail($input['user_id']);
-            CampUser::create([
-                'user_id' => $user->id,
-                'camp_id' => $camp->id,
-                'role_id' => $input['role_id'], ]
-            );
+
+        if (!Auth::user()->demo) {
+            if ($input['user_id']) {
+                $aktUser = Auth::user();
+                $camp = $aktUser->camp;
+                $user = User::findOrFail($input['user_id']);
+                CampUser::create([
+                        'user_id' => $user->id,
+                        'camp_id' => $camp->id,
+                        'role_id' => $input['role_id'],]
+                );
+            }
         }
 
         return redirect('/admin/users/create');
@@ -408,7 +421,10 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         //
-        User::findOrFail($id)->delete();
+
+        if (!Auth::user()->demo) {
+            User::findOrFail($id)->delete();
+        }
 
         return redirect('/admin/users');
     }
